Arrange condition util classes.
New KeystrokeTiming class for reusing.
Rename KeystrokeChecker to Keystroke.
Move Interval and KeystrokeChecker to new subpackage.
Move Interval to independent file.
Move KeystrokeChecker to independent file.
